make it work on 64 bit machines like Opteron
git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@404 d8f3d347-7192-4e35-a67b-6d701703d080

+++ tools/nosiverify/NosIVerify.c
@@ -51,10 +51,10 @@
 */
 typedef signed char  i8;
 typedef signed short i16;
-typedef signed long  i32;
+typedef signed int   i32;
 typedef unsigned char  u8;
 typedef unsigned short u16;
-typedef unsigned long  u32;
+typedef unsigned int   u32;
 #ifdef __GNUC__
 typedef long long i64;
 typedef unsigned long long u64;
